+++ libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: dacaball <dacaball@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 11:41:03 by dacaball          #+#    #+#             */
-/*   Updated: 2023/09/15 10:24:36 by dacaball         ###   ########.fr       */
+/*   Updated: 2023/09/15 11:25:35 by dacaball         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	int		j;
 
 	i = 0;
 	j = 0;
-	while ((big[i] != '\0') && ((unsigned long)i < len))
+	while ((big[i] != '\0') && (i < len))
 	{
 		while ((big[i + j] == little[j])
-			&& (little[j] != '\0') && ((unsigned long)i + j < len))
+			&& (little[j] != '\0') && (i + j < len))
 			j++;
 		if (little[j] == '\0')
 			return ((char *)big + i);
